Bug 25485: Fix libstdc++.so.6 compatibility for Tor Browser
We ship our version of libstdc++.so.6 for backward compatibility of Tor
Browser on older systems like Ubuntu 16.04. However, a newer version of
the library may be installed on the user's system and we should detect
and use that. This commit adds support for detecting the library using a
C++ program that uses the latest ABI; if the program executes, we should
use the system library, and if not (for the older systems), we add our
libstdc++.so.6 to LD_LIBRARY_PATH.